Corrected my error from 332b6c3.
Corrected my error from 332b6c3. What is actually true is that Drogon
users are free to run the application in the build/ directory if they
want.

In the docs, "./" is not meant to indicate application root. Instead,
it is meant to indicate the current working directory (which makes
more sense anyway).