In ENG-04-2, make code snippets match what the provided drogon_ctl command generates (#39)
* Updated ENG-04-2-Controller-HttpController demo_v1_User.* code snippets.

Updated ENG-04-2-Controller-HttpController demo_v1_User.* code snippets to
match the source code actually generated by drogon_ctl. This may not be
desirable though, since the generated code is slightly less readable
than the generated code, due to the lack of indentation in the
generated code. If so, please disregard this commit.